Update awsm_stubber.go
Currently if we use stubs in a single test func multiple times the callIndex will increase while the stubs decrease. If we run a for loop for the doing multiple tests, and stub twice for a loop, the call index will eventually become 3 while the stubs available is lesser then 3 throwing an error out.